검색 및 리뷰별점 기능 구현 - 동현

+++ BLB_Project/src/main/java/com/mig/blb/review/model/vo/Review.java
@@ -19,7 +19,7 @@ public class Review {
 	private String revTitle;			//	REV_TITLE	VARCHAR2(300 BYTE)
 	private String revContent;			//	REV_CONTENT	VARCHAR2(1500 BYTE)
 	private Timestamp revEnrollDate;	//	REV_ENROLL_DATE	DATE
-	private int revRating; 				//	REV_RATING	NUMBER
+	private float revRating; 				//	REV_RATING	NUMBER
 	private String revStatus;			//	REV_STATUS	CHAR(1 BYTE)
 	private String serialNo;			//	SERIAL_NO	VARCHAR2(30 BYTE)
 	private String memberId;			//	MEMBER_ID	VARCHAR2(20 BYTE)
